requirements.txt
flask==3.0.0
transformers==4.36.2
torch>=2.0.0
nltk==3.8.1
python-dotenv==1.0.0
flask-limiter==3.5.0
numpy==1.26.2
python-docx==0.8.11
python-pptx==0.6.21
PyPDF2==3.0.1
pdfminer.six==20221105
